sonic/pyproject.toml
[tool.poetry]
name = "agent-git"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
litellm = "^1.60.6"
pydantic = "^2.10.6"
numpy = "^2.2.2"
python-decouple = "^3.8"
pygithub = "^2.5.0"
scikit-learn = "^1.6.1"
mercury = "^2.4.3"
jp-proxy-widget = "^1.0.10"


[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
pytest = "^8.3.4"
pre-commit = "^4.1.0"
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
